move unitxt to own test script
